requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyYAML==6.0
